+++ samples/DLCBuilder/RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 - Using an arrangement without any notes or chords no longer causes an error.
 - Improvements to error handling.
+- Added an option to open the XML file from the arrangement context menu.
 - Updated the UI framework to the latest version.
 
 ## 3.4.0 (Released 2025-02-02)